+++ src/index.ts
@@ -1,4 +1,4 @@
-export const version = '2.4.26';
+export const version = '2.4.31';
 
 // G2 自定义能力透出
 import * as G2 from '@antv/g2';