+++ src/plots/scatter/types.ts
@@ -8,6 +8,7 @@ import {
   SizeAttr,
   StyleAttr,
   TextOption,
+  TextStyle,
 } from '../../types';
 
 interface Labels extends Omit<TextOption, 'position'> {
@@ -60,6 +61,21 @@ export interface RegressionLineOptions {
    * @description  [[0,0],[100,100]]
    */
   readonly algorithm?: Array<[number, number]> | ((data: any) => Array<[number, number]>);
+  /**
+   * @title 显示回归方程式
+   * @description 默认为不显示回归方程式
+   */
+  readonly showEquation?: boolean;
+  /**
+   * @title 自定义回归方程式
+   * @description 只有当自定义 algorithm 时生效
+   */
+  readonly equation?: string;
+  /**
+   * @title 回归线方程式样式
+   * @description 自定义文本样式，请参考 TextStyle 配置
+   */
+  readonly equationStyle?: TextStyle;
 }
 
 export interface ScatterOptions extends Options {